refactor: use client nlp service in workspace
